# Controllers/Api/ApiCarrinhoController.cs
using System.Security.Claims;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Microsoft.EntityFrameworkCore;
using WebApplication1.Data;
using WebApplication1.Dtos.Api;
using WebApplication1.Extensions;
using WebApplication1.Models;
using WebApplication1.Services.Checkout;

namespace WebApplication1.Controllers.Api;

[ApiController]
[Authorize(Roles = "Consumidor")]
[Route("api/carrinho")]
[Produces("application/json")]
public sealed class ApiCarrinhoController : ControllerBase
{
    private readonly ApplicationDbContext _context;
    private readonly ICheckoutFacade _checkoutFacade;

    public ApiCarrinhoController(ApplicationDbContext context, ICheckoutFacade checkoutFacade)
    {
        _context = context;
        _checkoutFacade = checkoutFacade;
    }

    /// <summary>
    /// Retorna o carrinho do consumidor logado.
    /// </summary>
    [HttpGet]
    [ProducesResponseType(typeof(CarrinhoResponse), StatusCodes.Status200OK)]
    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
    public ActionResult<CarrinhoResponse> Obter()
    {
        return Ok(CriarResposta(ObterCarrinho()));
    }

    /// <summary>
    /// Adiciona um item ao carrinho.
    /// </summary>
    /// <remarks>
    /// Exemplo de request: { "produtoId": 1, "quantidade": 2 }
    /// </remarks>
    [HttpPost("itens")]
    [ProducesResponseType(typeof(CarrinhoResponse), StatusCodes.Status200OK)]
    [ProducesResponseType(StatusCodes.Status400BadRequest)]
    [ProducesResponseType(StatusCodes.Status404NotFound)]
    public async Task<ActionResult<CarrinhoResponse>> Adicionar(CarrinhoAdicionarRequest request, CancellationToken cancellationToken)
    {
        var produto = await _context.Produtos.Include(p => p.Lojista).FirstOrDefaultAsync(p => p.Id == request.ProdutoId, cancellationToken);
        if (produto == null)
        {
            return NotFound();
        }

        if (produto.Estoque <= 0)
        {
            return BadRequest(new { mensagem = "Produto indisponivel." });
        }

        var carrinho = ObterCarrinho();
        var quantidade = Math.Max(1, request.Quantidade);
        var item = carrinho.FirstOrDefault(i => i.ProdutoId == produto.Id);

        if (item == null)
        {
            var percentualComissao = await _context.ComissoesCategoria
                .Where(c => c.Categoria == produto.Categoria)
                .Select(c => c.Percentual)
                .FirstOrDefaultAsync(cancellationToken);

            carrinho.Add(new CarrinhoItem
            {
                ProdutoId = produto.Id,
                Nome = produto.Nome,
                Descricao = produto.Descricao,
                ImagemUrl = produto.ImagemUrl,
                Categoria = produto.Categoria,
                Preco = produto.Preco,
                Quantidade = Math.Min(quantidade, produto.Estoque),
                LojistaId = produto.LojistaId,
                NomeLojista = produto.Lojista?.NomeFantasia ?? "Lojista parceiro",
                PercentualComissao = percentualComissao == 0 ? 12m : percentualComissao
            });
        }
        else
        {
            item.Quantidade = Math.Min(item.Quantidade + quantidade, produto.Estoque);
        }

        SalvarCarrinho(carrinho);
        return Ok(CriarResposta(carrinho));
    }

    /// <summary>
    /// Atualiza a quantidade de um item no carrinho.
    /// </summary>
    [HttpPut("itens/{produtoId:int}")]
    [ProducesResponseType(typeof(CarrinhoResponse), StatusCodes.Status200OK)]
    [ProducesResponseType(StatusCodes.Status404NotFound)]
    public async Task<ActionResult<CarrinhoResponse>> Atualizar(int produtoId, CarrinhoAtualizarRequest request, CancellationToken cancellationToken)
    {
        var carrinho = ObterCarrinho();
        var item = carrinho.FirstOrDefault(i => i.ProdutoId == produtoId);
        if (item == null)
        {
            return NotFound();
        }

        var estoque = await _context.Produtos
            .Where(p => p.Id == produtoId)
            .Select(p => p.Estoque)
            .FirstOrDefaultAsync(cancellationToken);

        if (request.Quantidade <= 0)
        {
            carrinho.Remove(item);
        }
        else
        {
            item.Quantidade = Math.Min(request.Quantidade, estoque);
        }

        SalvarCarrinho(carrinho);
        return Ok(CriarResposta(carrinho));
    }

    /// <summary>
    /// Remove um item do carrinho.
    /// </summary>
    [HttpDelete("itens/{produtoId:int}")]
    [ProducesResponseType(typeof(CarrinhoResponse), StatusCodes.Status200OK)]
    public ActionResult<CarrinhoResponse> Remover(int produtoId)
    {
        var carrinho = ObterCarrinho();
        carrinho.RemoveAll(i => i.ProdutoId == produtoId);
        SalvarCarrinho(carrinho);
        return Ok(CriarResposta(carrinho));
    }

    private CarrinhoResponse CriarResposta(List<CarrinhoItem> carrinho)
    {
        var frete = _checkoutFacade.CalcularFrete(carrinho, "01001000");
        var subtotal = carrinho.Sum(i => i.Subtotal);

        return new CarrinhoResponse(
            carrinho.Select(i => i.ToResponse()).ToList(),
            carrinho.Sum(i => i.Quantidade),
            subtotal,
            frete,
            subtotal + frete);
    }

    private string ObterEmail()
    {
        return User.FindFirstValue(ClaimTypes.Email) ?? User.Identity?.Name ?? "cliente@beautymarket.com";
    }

    private List<CarrinhoItem> ObterCarrinho()
    {
        return HttpContext.Session.GetObjectFromJson<List<CarrinhoItem>>($"Carrinho_{ObterEmail()}") ?? new List<CarrinhoItem>();
    }

    private void SalvarCarrinho(List<CarrinhoItem> carrinho)
    {
        HttpContext.Session.SetObjectAsJson($"Carrinho_{ObterEmail()}", carrinho);
    }
}