# docs/Entrega 4/api/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Beauty Marketplace API",
    "description": "Documentacao OpenAPI da Entrega 4: catalogo, carrinho, checkout, pedidos e IA.",
    "version": "v1"
  },
  "paths": {
    "/api/carrinho": {
      "get": {
        "tags": [
          "ApiCarrinho"
        ],
        "summary": "Retorna o carrinho do consumidor logado.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarrinhoResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/carrinho/itens": {
      "post": {
        "tags": [
          "ApiCarrinho"
        ],
        "summary": "Adiciona um item ao carrinho.",
        "description": "Exemplo de request: { \"produtoId\": 1, \"quantidade\": 2 }",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CarrinhoAdicionarRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CarrinhoAdicionarRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CarrinhoAdicionarRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarrinhoResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/carrinho/itens/{produtoId}": {
      "put": {
        "tags": [
          "ApiCarrinho"
        ],
        "summary": "Atualiza a quantidade de um item no carrinho.",
        "parameters": [
          {
            "name": "produtoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CarrinhoAtualizarRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CarrinhoAtualizarRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CarrinhoAtualizarRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarrinhoResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ApiCarrinho"
        ],
        "summary": "Remove um item do carrinho.",
        "parameters": [
          {
            "name": "produtoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarrinhoResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/pedidos": {
      "get": {
        "tags": [
          "ApiPedidos"
        ],
        "summary": "Lista os pedidos do consumidor logado.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PedidoResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/pedidos/{id}": {
      "get": {
        "tags": [
          "ApiPedidos"
        ],
        "summary": "Consulta um pedido especifico do consumidor logado.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PedidoResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CarrinhoAdicionarRequest": {
        "type": "object",
        "properties": {
          "produtoId": {
            "type": "integer",
            "format": "int32"
          },
          "quantidade": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "CarrinhoAtualizarRequest": {
        "type": "object",
        "properties": {
          "quantidade": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "CarrinhoItemResponse": {
        "type": "object",
        "properties": {
          "produtoId": {
            "type": "integer",
            "format": "int32"
          },
          "nome": {
            "type": "string",
            "nullable": true
          },
          "categoria": {
            "type": "string",
            "nullable": true
          },
          "preco": {
            "type": "number",
            "format": "double"
          },
          "quantidade": {
            "type": "integer",
            "format": "int32"
          },
          "subtotal": {
            "type": "number",
            "format": "double"
          },
          "lojistaId": {
            "type": "integer",
            "format": "int32"
          },
          "nomeLojista": {
            "type": "string",
            "nullable": true
          },
          "percentualComissao": {
            "type": "number",
            "format": "double"
          },
          "valorComissao": {
            "type": "number",
            "format": "double"
          },
          "valorRepasseLojista": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "CarrinhoResponse": {
        "type": "object",
        "properties": {
          "itens": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CarrinhoItemResponse"
            },
            "nullable": true
          },
          "quantidadeTotal": {
            "type": "integer",
            "format": "int32"
          },
          "subtotal": {
            "type": "number",
            "format": "double"
          },
          "freteEstimado": {
            "type": "number",
            "format": "double"
          },
          "totalEstimado": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "PedidoItemResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "produtoId": {
            "type": "integer",
            "format": "int32"
          },
          "lojistaId": {
            "type": "integer",
            "format": "int32"
          },
          "nomeProduto": {
            "type": "string",
            "nullable": true
          },
          "nomeLojista": {
            "type": "string",
            "nullable": true
          },
          "quantidade": {
            "type": "integer",
            "format": "int32"
          },
          "precoUnitario": {
            "type": "number",
            "format": "double"
          },
          "valorComissao": {
            "type": "number",
            "format": "double"
          },
          "valorRepasseLojista": {
            "type": "number",
            "format": "double"
          },
          "codigoRastreio": {
            "type": "string",
            "nullable": true
          },
          "statusEntrega": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PedidoResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "usuarioEmail": {
            "type": "string",
            "nullable": true
          },
          "criadoEm": {
            "type": "string",
            "format": "date-time"
          },
          "metodoPagamento": {
            "type": "string",
            "nullable": true
          },
          "cepEntrega": {
            "type": "string",
            "nullable": true
          },
          "subtotal": {
            "type": "number",
            "format": "double"
          },
          "frete": {
            "type": "number",
            "format": "double"
          },
          "total": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "itens": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PedidoItemResponse"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  },
...
}